+++ src/App.js
@@ -2,7 +2,7 @@ import "./App.css";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 
 import BackGround from "./views/BackGround";
-import Header from './views/Header';
+import Header from './views/Headers';
 import Main from "./views/Main";
 import Search from "./views/Search";
